+++ subprojects/core/src/main/java/hu/bme/mit/theta/core/type/bvtype/BvLitExpr.java
@@ -248,7 +248,7 @@ public final class BvLitExpr extends NullaryExpr<BvType> implements LitExpr<BvTy
     }
 
     public IntLitExpr toInt() {
-        return Int(BvUtils.bvLitExprToBigInteger(this).intValue());
+        return Int(BvUtils.bvLitExprToBigInteger(this));
     }
 
     @Override
